@@ -1,11 +1,11 @@
 {
 	"name": "starcitizenwiki/embedvideo",
-	"version": "3.2.8",
+	"version": "3.3.0",
 	"type": "mediawiki-extension",
 	"description": "Adds a parser function embedding video from popular sources.",
 	"license": "MIT",
 	"require": {
-		"php": ">=7.0.0",
+		"php": ">=8.0.0",
 		"composer/installers": ">=1.0.1"
 	},
 	"require-dev": {
