@@ -81,6 +81,27 @@ class EmbedVideo {
 	}
 
 	/**
+	 * {{#evu}} parser function that tries to extract the service from the host name
+	 *
+	 * @param Parser $parser
+	 * @param PPFrame $frame
+	 * @param array $args
+	 * @param bool $fromTag
+	 * @return array
+	 */
+	public static function parseEVU( Parser $parser, PPFrame $frame, array $args, bool $fromTag = false ): array {
+		$host = parse_url( $args[0] ?? '', PHP_URL_HOST );
+
+		if ( is_string( $host ) ) {
+			$host = explode( '.', trim( $host, 'w.' ) )[0] ?? null;
+		}
+
+		array_unshift( $args, $host );
+
+		return self::parseEV( $parser, $frame, $args, $fromTag );
+	}
+
+	/**
 	 * Parse the values input from the {{#ev}} parser function
 	 *
 	 * @param Parser $parser The active Parser instance
@@ -129,8 +150,8 @@ class EmbedVideo {
 	 * @param array $arguments Method arguments
 	 */
 	public static function __callStatic( $name, $arguments ) {
-		if ( strpos( $name, 'parseTag' ) !== 0 ) {
-			return;
+		if ( !str_starts_with( $name, 'parseTag' ) ) {
+			return '';
 		}
 
 		[
@@ -151,9 +172,7 @@ class EmbedVideo {
 	 * @return array
 	 */
 	public function output(): array {
-		[
-			'service' => $service,
-		] = $this->args;
+		$service = $this->args['service'] ?? null;
 
 		try {
 			$enabledServices = $this->config->get( 'EmbedVideoEnabledServices' ) ?? [];
@@ -212,16 +231,14 @@ class EmbedVideo {
 		];
 
 		if ( $fromTag === true ) {
+			$supportedArgs['service'] = $args['service'] ?? null;
+
 			return array_merge( $supportedArgs, $args );
 		}
 
 		$keys = array_keys( $supportedArgs );
 
-		if ( $fromTag ) {
-			$serviceName = $args['service'] ?? null;
-		} else {
-			$serviceName = array_shift( $args );
-		}
+		$serviceName = array_shift( $args );
 
 		$counter = 0;
 
@@ -233,7 +250,7 @@ class EmbedVideo {
 			$pair = [ $arg ];
 			// Only split arg if it is not an url and not urlArgs
 			// phpcs:ignore Generic.Files.LineLength.TooLong
-			if ( ( $keys[$counter] !== 'urlArgs' || strpos( $arg, 'urlArgs' ) !== false ) && preg_match( '/https?:/', $arg ) !== 1 ) {
+			if ( ( $keys[$counter] !== 'urlArgs' || str_contains( $arg, 'urlArgs' ) ) && preg_match( '/https?:/', $arg ) !== 1 ) {
 				$pair = explode( '=', $arg, 2 );
 			}
 			$pair = array_map( 'trim', $pair );
@@ -254,7 +271,7 @@ class EmbedVideo {
 			++$counter;
 		}
 
-		$supportedArgs['service'] = $serviceName;
+		$supportedArgs['service'] = $serviceName ?? false;
 
 		// An intentional weak check
 		if ( $supportedArgs['autoresize'] == true ) {
@@ -303,11 +320,15 @@ class EmbedVideo {
 			'title' => $title,
 		] = $this->args;
 
-		// I am not using $parser->parseWidthParam() since it can not handle height only.  Example: x100
-		if ( stripos( $dimensions, 'x' ) !== false ) {
-			$dimensions = strtolower( $dimensions );
-			[ $width, $height ] = explode( 'x', $dimensions );
-		} elseif ( is_numeric( $dimensions ) ) {
+		$rpl = fn( $input ) => preg_replace( '/[a-z]/i', '', (string)$input );
+
+		// Height only
+		if ( !empty( $dimensions ) && strtolower( $dimensions )[0] === 'x' ) {
+			$height = $dimensions;
+		// Width and height
+		} elseif ( preg_match( '/[0-9]+(?:px)?x[0-9]+(?:px)?/i', $dimensions ?? '' ) ) {
+			[ $width, $height ] = array_map( fn( $dim ) => $rpl( $dim ), array_filter( explode( 'x', $dimensions ) ) );
+		} elseif ( is_numeric( $rpl( $dimensions ) ) ) {
 			$width = $dimensions;
 		}
 
@@ -318,9 +339,8 @@ class EmbedVideo {
 		$this->service = EmbedServiceFactory::newFromName( $service, $id );
 
 		// Let the service automatically handle bad dimensional values.
-		$this->service->setWidth( $width );
-
-		$this->service->setHeight( $height );
+		$this->service->setWidth( $rpl( (string)$width ) );
+		$this->service->setHeight( $rpl( (string)$height ) );
 
 		if ( $this->config->get( 'EmbedVideoRequireConsent' ) === true ) {
 			$this->service->setUrlArgs( $this->service->getAutoplayParameter() );
