@@ -5,9 +5,9 @@ declare( strict_types=1 );
 namespace MediaWiki\Extension\EmbedVideo\Media\TransformOutput;
 
 use MediaWiki\Extension\EmbedVideo\EmbedService\EmbedHtmlFormatter;
+use MediaWiki\Extension\EmbedVideo\EmbedService\LocalVideo;
 
 class VideoEmbedTransformOutput extends VideoTransformOutput {
-
 	/**
 	 * Outputs 'embed like' local files
 	 * This essentially replaces the <iframe> element with the local <video> in question
@@ -17,7 +17,7 @@ class VideoEmbedTransformOutput extends VideoTransformOutput {
 	 * @return string HTML
 	 */
 	public function toHtml( $options = [] ): string {
-		$service = new FauxEmbedService( $this, $this->parameters );
+		$service = new LocalVideo( $this, $this->parameters );
 
 		return EmbedHtmlFormatter::toHtml( $service, [
 			'service' => 'local-embed',
